+++ MyMomii/MyMomii/UI/SettingViews/SettingNotiView.swift
@@ -6,70 +6,19 @@
 
 import SwiftUI
 
-class NotificationManager {
-    
-    // singleton으로 만들어서 한 번에 관리
-    static let instance = NotificationManager()
-    
-    // 알림 권한 Alert
-    func requestNotificationAuthorization() {
-        UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .sound]) {
-            (success, error) in
-                // 콘솔 확인용
-                if let error = error {
-                    print("ERROR: \(error)")
-                } else {
-                    print("SUCCESS")
-                }
-        }
-    }
-
-    // 오전 9시에 알림 보내기
-    func scheduleNotification() {
-        let content = UNMutableNotificationContent()
-        content.title = "생리 예정 알림"
-        content.body = "오늘은 생리 예정일 입니다."
-        content.sound = .default
-
-        var dateComponents = DateComponents()
-        dateComponents.hour = 17
-        dateComponents.minute = 40
-        // MARK: - 생리 예정일 날짜 dateFormatter 필요
-//        dateComponents.year =
-//        dateComponents.month =
-//        dateComponents.day =
-        
-        let trigger = UNCalendarNotificationTrigger(dateMatching: dateComponents, repeats: true)
-        let request = UNNotificationRequest(
-            identifier: UUID().uuidString,
-            content: content,
-            trigger: trigger) // trigger는 3가지 방법 가능 (time, calendar, location)
-
-        UNUserNotificationCenter.current().add(request) { error in
-            if let error = error {
-                print("ERROR: \(error.localizedDescription)")
-            } else {
-                print("SUCCESS")
-            }
-        }
-    }
-
-    // 알림 취소 & 한 번 본 알람 자동 삭제
-    func cancelNotification() {
-        UNUserNotificationCenter.current().removeAllPendingNotificationRequests()
-        UNUserNotificationCenter.current().removeAllDeliveredNotifications()
-        
-    }
-}
-
 struct SettingNotiView: View {
     @State var isMensToday = true
+//    @State var expect: String = "20231115"
+//    @Binding var eventsArray: [String]
     
     var body: some View {
         VStack(spacing: 0) {
             HStack {
                 Text("알림 설정")
                     .bold28Black400()
+//                    .onTapGesture {
+//                        NotificationManager.instance.scheduleNotification(expectedDate: expect)
+//                    }
                 Spacer()
             }
             .padding(EdgeInsets(top: 20, leading: 16, bottom: 20, trailing: 16))
@@ -89,9 +38,18 @@ struct SettingNotiView: View {
                 .padding(.bottom, 20)
                 VStack(spacing: 0) {
                     ToggleList(toggleTitle: "생리 예정 당일", toggleIsOn: $isMensToday)
-                        .onChange(of: isMensToday) {
+                    
+                    // MARK: - 뷰 연결되면 .onChange로 바꿀 예정 (eventsArray 받아올 예정 : 상단에 binding을 확인하라 & text에 온탭제스처 삭제하라)
+//                        .onChange(of: eventsArray) {
+//                            if eventsArray.count > 0 { // 이벤트 어레이에 값이 들어오면, 스케쥴링 해버림 (그때 날짜는 제일 첫번째)
+//                                NotificationManager.instance.scheduleNotification(expectedDate: eventsArray[0])
+//                            } else {
+//                                NotificationManager.instance.cancelNotification()
+//                            }
+//                        }
+                        .onAppear {
                             if isMensToday == true { // 토글 켜면, 노티피케이션 받을 수 있게 하고, 스케쥴링 해버림
-                                NotificationManager.instance.scheduleNotification()
+                                NotificationManager.instance.requestNotificationAuthorization()
                             } else {
                                 NotificationManager.instance.cancelNotification()
                             }
